+++ pyslvs_ui/entities/edit_point.py
@@ -18,10 +18,7 @@ from .edit_point_ui import Ui_Dialog
 
 class EditPointDialog(QDialog, Ui_Dialog):
 
-    """Option dialog.
-
-    Only edit the target path after closed.
-    """
+    """Option dialog."""
 
     def __init__(
         self,