+++ docs/changelog.md
@@ -9,6 +9,7 @@
 
 **Development**:
 
++ Create QRotatableView API.
 + Fix dependencies compatibility.
 
 ## Version 19.12